Refactor: 토큰으로 현재 사용자를 가져오는 로직 적용
Refactor: 실적 목표 조회 기능 수정

+++ src/main/java/com/clover/salad/goal/command/application/controller/GoalCommandController.java
@@ -7,11 +7,13 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.clover.salad.goal.command.application.dto.GoalDTO;
 import com.clover.salad.goal.command.application.service.GoalCommandService;
+import com.clover.salad.security.JwtUtil;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,43 +24,59 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class GoalCommandController {
 	private final GoalCommandService goalCommandService;
+	private final JwtUtil jwtUtil;
 	
 	/* 설명. 실적 목표 등록 */
 	@PostMapping("/register")
-	public ResponseEntity<String> registerGoal(@RequestBody List<GoalDTO> goalList) {
-		
-		/* TODO. 설명. 임시 사원 id(토큰에서 파싱) */
-		int tempEmployeeId = 1;
+	public ResponseEntity<String> registerGoal(
+		@RequestHeader("Authorization") String token,
+		@RequestBody List<GoalDTO> goalList
+	) {
+		String pureToken = token.replace("Bearer ", "");
+		String code = jwtUtil.getUsername(pureToken);
 		
 		try {
 			log.info("Registering goal");
-			goalCommandService.registerGoal(goalList, tempEmployeeId);
+			goalCommandService.registerGoal(goalList, code);
 		} catch (Exception e) {
 			return ResponseEntity.badRequest().body(e.getMessage());
 		}
 		return ResponseEntity.ok("Goal Registered");
 	}
 	
+	/* 설명. 실적 목표 수정 */
 	@PutMapping("/change")
-	public ResponseEntity<String> changeGoal(@RequestBody List<GoalDTO> goalList) {
+	public ResponseEntity<String> changeGoal(
+		@RequestHeader("Authorization") String token,
+		@RequestBody List<GoalDTO> goalList
+	) {
 		
-		/* TODO. 설명. 임시 사원 id(토큰에서 파싱) */
-		int tempEmployeeId = 1;
+		String pureToken = token.replace("Bearer ", "");
+		String code = jwtUtil.getUsername(pureToken);
 		
 		try {
 			log.info("Changing goal");
-			goalCommandService.changeGoal(goalList, tempEmployeeId);
+			goalCommandService.changeGoal(goalList, code);
 		} catch (Exception e) {
 			return ResponseEntity.badRequest().body(e.getMessage());
 		}
 		return ResponseEntity.ok("Goal Changed");
 	}
 	
-	/* 설명. 관리자가 퇴사 등으로 인해 달성할 수 없어진 목표를 삭제하는 기능
+	/* 설명. 실적 목표 삭제
+	 *  관리자가 퇴사 등으로 인해 달성할 수 없어진 목표를 삭제하는 기능
 	 *  목표를 조회 후 선택해서 삭제
 	 * */
 	@DeleteMapping("/delete")
-	public ResponseEntity<String> deleteGoal(@RequestBody List<GoalDTO> goalList) {
+	public ResponseEntity<String> deleteGoal(
+		@RequestHeader("Authorization") String token,
+		@RequestBody List<GoalDTO> goalList
+	) {
+		String pureToken = token.replace("Bearer ", "");
+		if(!jwtUtil.getAuthentication(pureToken).getAuthorities().toString().equals("ROLE_ADMIN")) {
+			return ResponseEntity.badRequest().body("관리자만 목표를 삭제할 수 있습니다");
+		}
+		
 		try {
 			log.info("Deleting goal");
 			goalCommandService.deleteGoal(goalList);
